requirements.txt
playwright>=1.40.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
icalendar>=5.0.0
recurring-ical-events>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
